resume_screening_automation/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
groq
python-dotenv
python-docx
pdfplumber
python-multipart
email-validator
alembic
pytest
